rag-advanced/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aioprocessing==2.0.1
aiosignal==1.3.1
analytics-python==1.2.9
annotated-types==0.7.0
anyio==4.6.1
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
bm25s==0.2.2
cbor==1.0.0
cbor2==5.6.5
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
contourpy==1.3.0
cramjam==2.8.4
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.16
emoji==2.14.0
executing==2.1.0
fastparquet==2024.5.0
fasttext==0.9.3
fasttext-langdetect==1.0.5
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
gql==3.5.0
graphql-core==3.2.4
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.10
ijson==3.3.0
importlib_metadata==8.5.0
inscriptis==2.5.0
instructor==1.5.2
ipykernel==6.29.5
ipython==8.28.0
ipywidgets==8.1.5
ir_datasets==0.5.8
janus==1.0.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
Levenshtein==0.26.0
litellm==1.49.2
llvmlite==0.43.0
loguru==0.7.2
lxml==5.3.0
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.1.0
nest-asyncio==1.6.0
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
openai==1.51.2
orjson==3.10.7
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==16.1.0
pybind11==2.13.6
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pymdown-extensions==10.11.2
pyparsing==3.2.0
PyPDF2==3.0.1
PyStemmer==2.2.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-frontmatter==1.1.0
python-json-logger==2.0.7
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
ranx==0.3.20
RapidFuzz==3.10.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.2
rouge==1.0.1
rpds-py==0.20.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
sentry-sdk==2.16.0
set-env-colab-kaggle-dotenv==0.1.3
setproctitle==1.3.3
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
syncer==2.0.3
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
trec-car-tools==2.6
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
unlzw3==0.2.2
urllib3==2.2.3
uuid_utils==0.9.0
wandb==0.18.3
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
wcwidth==0.2.13
weave==0.51.12
Werkzeug==3.0.4
widgetsnbextension==4.0.13
yarl==1.15.1
zipp==3.20.2
zlib-state==0.1.9
